+++ packages/components/src/circular-option-picker/index.tsx
@@ -6,21 +6,30 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
+import { useInstanceId } from '@wordpress/compose';
+import { createContext, useContext, useEffect } from '@wordpress/element';
 import { Icon, check } from '@wordpress/icons';
+import { isRTL } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import Button from '../button';
+import { Composite, CompositeItem, useCompositeState } from '../composite';
 import Dropdown from '../dropdown';
 import Tooltip from '../tooltip';
 import type {
 	CircularOptionPickerProps,
 	DropdownLinkActionProps,
+	OptionGroupProps,
 	OptionProps,
 } from './types';
 import type { WordPressComponentProps } from '../ui/context';
 import type { ButtonAsButtonProps } from '../button/types';
+
+const CircularOptionPickerContext = createContext( {} );
+
+const hasSelectedOption = new Map();
 
 export function Option( {
 	className,
@@ -29,13 +38,44 @@ export function Option( {
 	tooltipText,
 	...additionalProps
 }: OptionProps ) {
-	const optionButton = (
-		<Button
-			isPressed={ isSelected }
-			className="components-circular-option-picker__option"
+	const compositeState = useContext( CircularOptionPickerContext );
+	const {
+		baseId = 'option',
+		currentId,
+		setCurrentId,
+	} = compositeState as any;
+	const id = useInstanceId( Option, baseId );
+
+	useEffect( () => {
+		// If we call `setCurrentId` here, it doesn't update for other
+		// Option renders in the same pass. So we have to store our own
+		// map to make sure that we only set the first selected option.
+		// We still need to check `currentId` because the control will
+		// update this as the user moves around, and that state should
+		// be maintained as the group gains and loses focus.
+		if ( isSelected && ! currentId && ! hasSelectedOption.get( baseId ) ) {
+			hasSelectedOption.set( baseId, true );
+			setCurrentId( id );
+		}
+	} );
+
+	const optionControl = (
+		<CompositeItem
+			as={ Button }
+			className={ classnames(
+				'components-circular-option-picker__option',
+				{
+					'is-pressed': isSelected,
+				}
+			) }
+			id={ id }
 			{ ...additionalProps }
+			{ ...compositeState }
+			role="option"
+			aria-selected={ !! isSelected }
 		/>
 	);
+
 	return (
 		<div
 			className={ classnames(
@@ -44,9 +84,9 @@ export function Option( {
 			) }
 		>
 			{ tooltipText ? (
-				<Tooltip text={ tooltipText }>{ optionButton }</Tooltip>
+				<Tooltip text={ tooltipText }>{ optionControl }</Tooltip>
 			) : (
-				optionButton
+				optionControl
 			) }
 			{ isSelected && (
 				<Icon
@@ -54,6 +94,30 @@ export function Option( {
 					{ ...( selectedIconProps ? selectedIconProps : {} ) }
 				/>
 			) }
+		</div>
+	);
+}
+
+export function OptionGroup( {
+	className,
+	options,
+	...additionalProps
+}: OptionGroupProps ) {
+	const { 'aria-label': ariaLabel, 'aria-labelledby': ariaLabelledby } =
+		additionalProps as any;
+	const role = ariaLabel || ariaLabelledby ? 'group' : undefined;
+
+	return (
+		<div
+			{ ...additionalProps }
+			role={ role }
+			className={ classnames(
+				'components-circular-option-picker__option-group',
+				'components-circular-option-picker__swatches',
+				className
+			) }
+		>
+			{ options }
 		</div>
 	);
 }
@@ -152,28 +216,48 @@ export function ButtonAction( {
  */
 
 function CircularOptionPicker( props: CircularOptionPickerProps ) {
-	const { actions, className, options, children } = props;
+	const {
+		actions,
+		className,
+		id: idProp,
+		options,
+		children,
+		loop = true,
+		...additionalProps
+	} = props;
+	const id = useInstanceId( CircularOptionPicker, 'option-picker', idProp );
+	const rtl = isRTL();
+	const compositeState = useCompositeState( { baseId: id, loop, rtl } );
+
 	return (
-		<div
+		<Composite
+			{ ...additionalProps }
+			{ ...compositeState }
+			role={ 'listbox' }
 			className={ classnames(
 				'components-circular-option-picker',
 				className
 			) }
 		>
-			<div className="components-circular-option-picker__swatches">
-				{ options }
-			</div>
-			{ children }
-			{ actions && (
-				<div className="components-circular-option-picker__custom-clear-wrapper">
-					{ actions }
+			<CircularOptionPickerContext.Provider value={ compositeState }>
+				<div
+					className={ 'components-circular-option-picker__swatches' }
+				>
+					{ options }
 				</div>
-			) }
-		</div>
+				{ children }
+				{ actions && (
+					<div className="components-circular-option-picker__custom-clear-wrapper">
+						{ actions }
+					</div>
+				) }
+			</CircularOptionPickerContext.Provider>
+		</Composite>
 	);
 }
 
 CircularOptionPicker.Option = Option;
+CircularOptionPicker.OptionGroup = OptionGroup;
 CircularOptionPicker.ButtonAction = ButtonAction;
 CircularOptionPicker.DropdownLinkAction = DropdownLinkAction;